+++ tests/Integration/Services/CRM/Quote/Service/QuoteContactTest.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Vadim Soluyanov <vadimsallee@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -11,13 +11,12 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Company\Service;
+namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service;
 
 
 use Bitrix24\SDK\Services\CRM\Common\ContactConnection;
-use Bitrix24\SDK\Services\CRM\Company\Service\CompanyContact;
+use Bitrix24\SDK\Services\CRM\Quote\Service\QuoteContact;
 use Bitrix24\SDK\Services\ServiceBuilder;
-use Bitrix24\SDK\Tests\Builders\Services\CRM\CompanyBuilder;
 use Bitrix24\SDK\Tests\Builders\Services\CRM\ContactBuilder;
 use Bitrix24\SDK\Tests\Integration\Fabric;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -26,20 +25,20 @@ use PHPUnit\Framework\TestCase;
 use Bitrix24\SDK\Core;
 use Bitrix24\SDK\Tests\CustomAssertions\CustomBitrix24Assertions;
 
-#[CoversClass(CompanyContact::class)]
-#[CoversMethod(CompanyContact::class, 'fields')]
-#[CoversMethod(CompanyContact::class, 'setItems')]
-#[CoversMethod(CompanyContact::class, 'deleteItems')]
-#[CoversMethod(CompanyContact::class, 'get')]
-#[CoversMethod(CompanyContact::class, 'add')]
-#[CoversMethod(CompanyContact::class, 'delete')]
-class CompanyContactTest extends TestCase
+#[CoversClass(QuoteContact::class)]
+#[CoversMethod(QuoteContact::class, 'fields')]
+#[CoversMethod(QuoteContact::class, 'setItems')]
+#[CoversMethod(QuoteContact::class, 'deleteItems')]
+#[CoversMethod(QuoteContact::class, 'get')]
+#[CoversMethod(QuoteContact::class, 'add')]
+#[CoversMethod(QuoteContact::class, 'delete')]
+class QuoteContactTest extends TestCase
 {
     use CustomBitrix24Assertions;
 
     private ServiceBuilder $sb;
 
-    private array $createdCompanies = [];
+    private array $createdQuotes = [];
     private array $createdContacts = [];
 
     public function setUp(): void
@@ -49,7 +48,7 @@ class CompanyContactTest extends TestCase
 
     public function tearDown(): void
     {
-        foreach ($this->sb->getCRMScope()->company()->batch->delete($this->createdCompanies) as $result) {
+        foreach ($this->sb->getCRMScope()->quote()->batch->delete($this->createdQuotes) as $result) {
         }
 
         foreach ($this->sb->getCRMScope()->contact()->batch->delete($this->createdContacts) as $result) {
@@ -58,21 +57,21 @@ class CompanyContactTest extends TestCase
 
     public function testSet(): void
     {
-        $companyId = $this->sb->getCRMScope()->company()->add((new CompanyBuilder())->build())->getId();
-        $this->createdCompanies[] = $companyId;
+        $quoteId = $this->sb->getCRMScope()->quote()->add(['TITLE' => 'test quote'])->getId();
+        $this->createdQuotes[] = $quoteId;
 
         $contactIdOne = $this->sb->getCRMScope()->contact()->add((new ContactBuilder())->build())->getId();
         $this->createdContacts[] = $contactIdOne;
         $contactIdTwo = $this->sb->getCRMScope()->contact()->add((new ContactBuilder())->build())->getId();
         $this->createdContacts[] = $contactIdTwo;
 
-        $this->sb->getCRMScope()->companyContact()->setItems($companyId, [
+        $this->sb->getCRMScope()->quoteContact()->setItems($quoteId, [
             new ContactConnection($contactIdOne, 100, true),
             new ContactConnection($contactIdTwo, 100, false),
         ]);
 
         $connectedId = [$contactIdOne, $contactIdTwo];
-        $connectedContacts = $this->sb->getCRMScope()->companyContact()->get($companyId)->getContactConnections();
+        $connectedContacts = $this->sb->getCRMScope()->quoteContact()->get($quoteId)->getContactConnections();
 
         foreach ($connectedContacts as $item) {
             $this->assertContains($item->CONTACT_ID, $connectedId);
@@ -81,15 +80,15 @@ class CompanyContactTest extends TestCase
 
     public function testAdd(): void
     {
-        $companyId = $this->sb->getCRMScope()->company()->add((new CompanyBuilder())->build())->getId();
-        $this->createdCompanies[] = $companyId;
+        $quoteId = $this->sb->getCRMScope()->quote()->add(['TITLE' => 'test quote'])->getId();
+        $this->createdQuotes[] = $quoteId;
 
         $contactIdOne = $this->sb->getCRMScope()->contact()->add((new ContactBuilder())->build())->getId();
         $this->createdContacts[] = $contactIdOne;
 
         $this->assertTrue(
-            $this->sb->getCRMScope()->companyContact()->add(
-                $companyId,
+            $this->sb->getCRMScope()->quoteContact()->add(
+                $quoteId,
                 new ContactConnection($contactIdOne, 100, true)
             )->isSuccess()
         );
@@ -98,14 +97,14 @@ class CompanyContactTest extends TestCase
         $this->createdContacts[] = $contactIdTwo;
 
         $this->assertTrue(
-            $this->sb->getCRMScope()->companyContact()->add(
-                $companyId,
+            $this->sb->getCRMScope()->quoteContact()->add(
+                $quoteId,
                 new ContactConnection($contactIdTwo, 100, true)
             )->isSuccess()
         );
 
         $connectedId = [$contactIdOne, $contactIdTwo];
-        $connectedContacts = $this->sb->getCRMScope()->companyContact()->get($companyId)->getContactConnections();
+        $connectedContacts = $this->sb->getCRMScope()->quoteContact()->get($quoteId)->getContactConnections();
 
         foreach ($connectedContacts as $item) {
             $this->assertContains($item->CONTACT_ID, $connectedId);
@@ -114,8 +113,8 @@ class CompanyContactTest extends TestCase
 
     public function testDeleteItems(): void
     {
-        $companyId = $this->sb->getCRMScope()->company()->add((new CompanyBuilder())->build())->getId();
-        $this->createdCompanies[] = $companyId;
+        $quoteId = $this->sb->getCRMScope()->quote()->add(['TITLE' => 'test quote'])->getId();
+        $this->createdQuotes[] = $quoteId;
 
         $contactIdOne = $this->sb->getCRMScope()->contact()->add((new ContactBuilder())->build())->getId();
         $this->createdContacts[] = $contactIdOne;
@@ -123,21 +122,21 @@ class CompanyContactTest extends TestCase
         $this->createdContacts[] = $contactIdTwo;
 
         $this->assertTrue(
-            $this->sb->getCRMScope()->companyContact()->setItems($companyId, [
+            $this->sb->getCRMScope()->quoteContact()->setItems($quoteId, [
                 new ContactConnection($contactIdOne, 100, true),
                 new ContactConnection($contactIdTwo, 100, false),
             ])->isSuccess()
         );
 
-        $this->assertTrue($this->sb->getCRMScope()->companyContact()->deleteItems($companyId)->isSuccess());
+        $this->assertTrue($this->sb->getCRMScope()->quoteContact()->deleteItems($quoteId)->isSuccess());
 
-        $this->assertCount(0, $this->sb->getCRMScope()->companyContact()->get($companyId)->getContactConnections());
+        $this->assertCount(0, $this->sb->getCRMScope()->quoteContact()->get($quoteId)->getContactConnections());
     }
 
     public function testDelete(): void
     {
-        $companyId = $this->sb->getCRMScope()->company()->add((new CompanyBuilder())->build())->getId();
-        $this->createdCompanies[] = $companyId;
+        $quoteId = $this->sb->getCRMScope()->quote()->add(['TITLE' => 'test quote'])->getId();
+        $this->createdQuotes[] = $quoteId;
 
         $contactIdOne = $this->sb->getCRMScope()->contact()->add((new ContactBuilder())->build())->getId();
         $this->createdContacts[] = $contactIdOne;
@@ -145,27 +144,27 @@ class CompanyContactTest extends TestCase
         $this->createdContacts[] = $contactIdTwo;
 
         $this->assertTrue(
-            $this->sb->getCRMScope()->companyContact()->setItems($companyId, [
+            $this->sb->getCRMScope()->quoteContact()->setItems($quoteId, [
                 new ContactConnection($contactIdOne, 100, true),
                 new ContactConnection($contactIdTwo, 100, false),
             ])->isSuccess()
         );
 
-        $this->assertTrue($this->sb->getCRMScope()->companyContact()->delete($companyId, $contactIdTwo)->isSuccess());
+        $this->assertTrue($this->sb->getCRMScope()->quoteContact()->delete($quoteId, $contactIdTwo)->isSuccess());
 
-        $this->assertCount(1, $this->sb->getCRMScope()->companyContact()->get($companyId)->getContactConnections());
+        $this->assertCount(1, $this->sb->getCRMScope()->quoteContact()->get($quoteId)->getContactConnections());
     }
 
     public function testSetWithEmptyConnections(): void
     {
         $this->expectException(Core\Exceptions\InvalidArgumentException::class);
-        $this->sb->getCRMScope()->companyContact()->setItems(1, []);
+        $this->sb->getCRMScope()->quoteContact()->setItems(1, []);
     }
 
     public function testSetWithWrongType(): void
     {
         $this->expectException(Core\Exceptions\InvalidArgumentException::class);
-        /** @phpstan-ignore */
-        $this->sb->getCRMScope()->companyContact()->setItems(1, [new \DateTime()]);
+        /** @phpstan-ignore-next-line */
+        $this->sb->getCRMScope()->quoteContact()->setItems(1, [new \DateTime()]);
     }
 }