+++ tests/Integration/Services/CRM/Quote/Service/QuoteProductRowsTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Lead\Service;
+namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service;
 
 use Money\Currencies\ISOCurrencies;
 use Money\Currency;
@@ -20,31 +20,39 @@ use Money\Money;
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Exceptions\TransportException;
 use Bitrix24\SDK\Services\CRM\Common\Result\DiscountType;
-use Bitrix24\SDK\Services\CRM\Lead\Result\LeadProductRowItemResult;
-use Bitrix24\SDK\Services\CRM\Lead\Service\Lead;
-use Bitrix24\SDK\Services\CRM\Lead\Service\LeadProductRows;
+use Bitrix24\SDK\Services\CRM\Quote\Result\QuoteProductRowItemResult;
+use Bitrix24\SDK\Services\CRM\Quote\Service\Quote;
+use Bitrix24\SDK\Services\CRM\Quote\Service\QuoteProductRows;
 use Bitrix24\SDK\Tests\Builders\DemoDataGenerator;
 use Bitrix24\SDK\Tests\Integration\Fabric;
 use MoneyPHP\Percentage\Percentage;
 use PHPUnit\Framework\TestCase;
 use Typhoon\Reflection\TyphoonReflector;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Lead\Service\LeadProductRows::class)]
-class LeadProductRowsTest extends TestCase
+#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Quote\Service\QuoteProductRows::class)]
+class QuoteProductRowsTest extends TestCase
 {
-    private Lead $leadService;
+    private Quote $quoteService;
 
-    private LeadProductRows $leadProductRowsService;
+    private QuoteProductRows $quoteProductRowsService;
 
     private DecimalMoneyFormatter $decimalMoneyFormatter;
 
     private TyphoonReflector $typhoonReflector;
+    
+    protected function setUp(): void
+    {
+        $this->quoteService = Fabric::getServiceBuilder()->getCRMScope()->quote();
+        $this->quoteProductRowsService = Fabric::getServiceBuilder()->getCRMScope()->quoteProductRows();
+        $this->decimalMoneyFormatter = new DecimalMoneyFormatter(new ISOCurrencies());
+        $this->typhoonReflector = TyphoonReflector::build();
+    }
 
     public function testAllSystemPropertiesAnnotated(): void
     {
-        $leadId = $this->leadService->add(['TITLE' => 'test lead'])->getId();
-        $this->leadProductRowsService->set(
-            $leadId,
+        $quoteId = $this->quoteService->add(['TITLE' => 'test quote'])->getId();
+        $this->quoteProductRowsService->set(
+            $quoteId,
             [
                 [
                     'PRODUCT_NAME' => sprintf('product name %s', time()),
@@ -53,9 +61,9 @@ class LeadProductRowsTest extends TestCase
             ]
         );
         // get response from server with actual keys
-        $propListFromApi = array_keys($this->leadProductRowsService->get($leadId)->getCoreResponse()->getResponseData()->getResult()['result']['rows'][0]);
+        $propListFromApi = array_keys($this->quoteProductRowsService->get($quoteId)->getCoreResponse()->getResponseData()->getResult()['result']['rows'][0]);
         // parse keys from phpdoc annotation
-        $collection = $this->typhoonReflector->reflectClass(LeadProductRowItemResult::class)->properties();
+        $collection = $this->typhoonReflector->reflectClass(QuoteProductRowItemResult::class)->properties();
         $propsFromAnnotations = [];
         foreach ($collection as $meta) {
             if ($meta->isAnnotated() && !$meta->isNative()) {
@@ -65,11 +73,11 @@ class LeadProductRowsTest extends TestCase
 
         $this->assertEquals($propListFromApi, $propsFromAnnotations,
             sprintf('in phpdocs annotations for class %s cant find fields from actual api response: %s',
-                LeadProductRowItemResult::class,
+                QuoteProductRowItemResult::class,
                 implode(', ', array_values(array_diff($propListFromApi, $propsFromAnnotations)))
             ));
         
-        $this->leadService->delete($leadId);
+        $this->quoteService->delete($quoteId);
     }
 
     /**
@@ -78,13 +86,13 @@ class LeadProductRowsTest extends TestCase
      */
     public function testSet(): void
     {
-        $leadId = $this->leadService->add(['TITLE' => sprintf('test lead %s', time())])->getId();
-        $lead = $this->leadService->get($leadId)->lead();
-        $price = new Money(100000, $lead->CURRENCY_ID);
-        $discount = new Money(50012, $lead->CURRENCY_ID);
+        $quoteId = $this->quoteService->add(['TITLE' => sprintf('test quote %s', time())])->getId();
+        $quote = $this->quoteService->get($quoteId)->quote();
+        $price = new Money(100000, $quote->CURRENCY_ID);
+        $discount = new Money(50012, $quote->CURRENCY_ID);
         $this::assertTrue(
-            $this->leadProductRowsService->set(
-                $leadId,
+            $this->quoteProductRowsService->set(
+                $quoteId,
                 [
                     [
                         'PRODUCT_NAME' => sprintf('product name %s', time()),
@@ -95,9 +103,9 @@ class LeadProductRowsTest extends TestCase
                 ]
             )->isSuccess()
         );
-        $leadProductRowItemsResult = $this->leadProductRowsService->get($leadId);
-        $this->assertCount(1, $leadProductRowItemsResult->getProductRows());
-        $productRow = $leadProductRowItemsResult->getProductRows()[0];
+        $quoteProductRowItemsResult = $this->quoteProductRowsService->get($quoteId);
+        $this->assertCount(1, $quoteProductRowItemsResult->getProductRows());
+        $productRow = $quoteProductRowItemsResult->getProductRows()[0];
         $this->assertEquals($price, $productRow->PRICE);
         $this->assertEquals(DiscountType::monetary, $productRow->DISCOUNT_TYPE_ID);
         $this->assertEquals($discount, $productRow->DISCOUNT_SUM);
@@ -105,18 +113,18 @@ class LeadProductRowsTest extends TestCase
         $calculatedPercentage = new Percentage((string)((int)$discount->getAmount() / 100));
         $this->assertEquals($calculatedPercentage, $productRow->DISCOUNT_RATE);
         
-        $this->leadService->delete($leadId);
+        $this->quoteService->delete($quoteId);
     }
 
     public function testGet(): void
     {
-        $leadId = $this->leadService->add(['TITLE' => sprintf('test lead %s', time())])->getId();
-        $lead = $this->leadService->get($leadId)->lead();
-        $price = new Money(100000, $lead->CURRENCY_ID);
-        $discount = new Money(0, $lead->CURRENCY_ID);
+        $quoteId = $this->quoteService->add(['TITLE' => sprintf('test quote %s', time())])->getId();
+        $quote = $this->quoteService->get($quoteId)->quote();
+        $price = new Money(100000, $quote->CURRENCY_ID);
+        $discount = new Money(0, $quote->CURRENCY_ID);
         $this::assertTrue(
-            $this->leadProductRowsService->set(
-                $leadId,
+            $this->quoteProductRowsService->set(
+                $quoteId,
                 [
                     [
                         'PRODUCT_NAME' => sprintf('product name %s', time()),
@@ -125,22 +133,15 @@ class LeadProductRowsTest extends TestCase
                 ]
             )->isSuccess()
         );
-        $leadProductRowItemsResult = $this->leadProductRowsService->get($leadId);
-        $this->assertCount(1, $leadProductRowItemsResult->getProductRows());
-        $productRow = $leadProductRowItemsResult->getProductRows()[0];
+        $quoteProductRowItemsResult = $this->quoteProductRowsService->get($quoteId);
+        $this->assertCount(1, $quoteProductRowItemsResult->getProductRows());
+        $productRow = $quoteProductRowItemsResult->getProductRows()[0];
         $this->assertEquals($price, $productRow->PRICE);
         $this->assertEquals(DiscountType::percentage, $productRow->DISCOUNT_TYPE_ID);
         $this->assertEquals($discount, $productRow->DISCOUNT_SUM);
         $this->assertEquals(Percentage::zero(), $productRow->DISCOUNT_RATE);
         
-        $this->leadService->delete($leadId);
+        $this->quoteService->delete($quoteId);
     }
 
-    protected function setUp(): void
-    {
-        $this->leadService = Fabric::getServiceBuilder()->getCRMScope()->lead();
-        $this->leadProductRowsService = Fabric::getServiceBuilder()->getCRMScope()->leadProductRows();
-        $this->decimalMoneyFormatter = new DecimalMoneyFormatter(new ISOCurrencies());
-        $this->typhoonReflector = TyphoonReflector::build();
-    }
 }