+++ tests/Integration/Services/CRM/Quote/Service/QuoteTest.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Vadim Soluyanov <vadimsallee@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -11,14 +11,13 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Lead\Service;
+namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Exceptions\TransportException;
 use Bitrix24\SDK\Core;
-use Bitrix24\SDK\Services\CRM\Contact\Result\ContactItemResult;
-use Bitrix24\SDK\Services\CRM\Lead\Result\LeadItemResult;
-use Bitrix24\SDK\Services\CRM\Lead\Service\Lead;
+use Bitrix24\SDK\Services\CRM\Quote\Result\QuoteItemResult;
+use Bitrix24\SDK\Services\CRM\Quote\Service\Quote;
 use Bitrix24\SDK\Tests\CustomAssertions\CustomBitrix24Assertions;
 use Bitrix24\SDK\Tests\Integration\Fabric;
 use PHPUnit\Framework\Attributes\CoversFunction;
@@ -26,37 +25,42 @@ use PHPUnit\Framework\Attributes\CoversMethod;
 use PHPUnit\Framework\TestCase;
 
 /**
- * Class LeadTest
+ * Class QuoteTest
  *
- * @package Bitrix24\SDK\Tests\Integration\Services\CRM\Lead\Service
+ * @package Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service
  */
-#[CoversMethod(Lead::class,'add')]
-#[CoversMethod(Lead::class,'delete')]
-#[CoversMethod(Lead::class,'get')]
-#[CoversMethod(Lead::class,'list')]
-#[CoversMethod(Lead::class,'fields')]
-#[CoversMethod(Lead::class,'update')]
-#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Lead\Service\Lead::class)]
-class LeadTest extends TestCase
+#[CoversMethod(Quote::class,'add')]
+#[CoversMethod(Quote::class,'delete')]
+#[CoversMethod(Quote::class,'get')]
+#[CoversMethod(Quote::class,'list')]
+#[CoversMethod(Quote::class,'fields')]
+#[CoversMethod(Quote::class,'update')]
+#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Quote\Service\Quote::class)]
+class QuoteTest extends TestCase
 {
     use CustomBitrix24Assertions;
-    protected Lead $leadService;
+    protected Quote $quoteService;
+    
+    protected function setUp(): void
+    {
+        $this->quoteService = Fabric::getServiceBuilder()->getCRMScope()->quote();
+    }
 
     public function testAllSystemFieldsAnnotated(): void
     {
-        $propListFromApi = (new Core\Fields\FieldsFilter())->filterSystemFields(array_keys($this->leadService->fields()->getFieldsDescription()));
-        $this->assertBitrix24AllResultItemFieldsAnnotated($propListFromApi, LeadItemResult::class);
+        $propListFromApi = (new Core\Fields\FieldsFilter())->filterSystemFields(array_keys($this->quoteService->fields()->getFieldsDescription()));
+        $this->assertBitrix24AllResultItemFieldsAnnotated($propListFromApi, QuoteItemResult::class);
     }
 
     public function testAllSystemFieldsHasValidTypeAnnotation():void
     {
-        $allFields = $this->leadService->fields()->getFieldsDescription();
+        $allFields = $this->quoteService->fields()->getFieldsDescription();
         $systemFieldsCodes = (new Core\Fields\FieldsFilter())->filterSystemFields(array_keys($allFields));
         $systemFields = array_filter($allFields, static fn($code): bool => in_array($code, $systemFieldsCodes, true), ARRAY_FILTER_USE_KEY);
 
         $this->assertBitrix24AllResultItemFieldsHasValidTypeAnnotation(
             $systemFields,
-            LeadItemResult::class);
+            QuoteItemResult::class);
     }
 
     /**
@@ -65,7 +69,7 @@ class LeadTest extends TestCase
      */
     public function testAdd(): void
     {
-        self::assertGreaterThan(1, $this->leadService->add(['TITLE' => 'test lead'])->getId());
+        self::assertGreaterThan(1, $this->quoteService->add(['TITLE' => 'test quote'])->getId());
     }
 
     /**
@@ -74,7 +78,7 @@ class LeadTest extends TestCase
      */
     public function testDelete(): void
     {
-        self::assertTrue($this->leadService->delete($this->leadService->add(['TITLE' => 'test lead'])->getId())->isSuccess());
+        self::assertTrue($this->quoteService->delete($this->quoteService->add(['TITLE' => 'test quote 1'])->getId())->isSuccess());
     }
 
     /**
@@ -83,7 +87,7 @@ class LeadTest extends TestCase
      */
     public function testFields(): void
     {
-        self::assertIsArray($this->leadService->fields()->getFieldsDescription());
+        self::assertIsArray($this->quoteService->fields()->getFieldsDescription());
     }
 
     /**
@@ -94,7 +98,7 @@ class LeadTest extends TestCase
     {
         self::assertGreaterThan(
             1,
-            $this->leadService->get($this->leadService->add(['TITLE' => 'test Lead'])->getId())->lead()->ID
+            $this->quoteService->get($this->quoteService->add(['TITLE' => 'test Quote 2'])->getId())->quote()->ID
         );
     }
 
@@ -104,8 +108,8 @@ class LeadTest extends TestCase
      */
     public function testList(): void
     {
-        $this->leadService->add(['TITLE' => 'test']);
-        self::assertGreaterThanOrEqual(1, $this->leadService->list([], [], ['ID', 'TITLE'])->getLeads());
+        $this->quoteService->add(['TITLE' => 'test']);
+        self::assertGreaterThanOrEqual(1, $this->quoteService->list([], [], ['ID', 'TITLE'])->getQuotes());
     }
 
     /**
@@ -114,11 +118,11 @@ class LeadTest extends TestCase
      */
     public function testUpdate(): void
     {
-        $addedItemResult = $this->leadService->add(['TITLE' => 'test lead']);
+        $addedItemResult = $this->quoteService->add(['TITLE' => 'test quote']);
         $newTitle = 'test2';
 
-        self::assertTrue($this->leadService->update($addedItemResult->getId(), ['TITLE' => $newTitle], [])->isSuccess());
-        self::assertEquals($newTitle, $this->leadService->get($addedItemResult->getId())->lead()->TITLE);
+        self::assertTrue($this->quoteService->update($addedItemResult->getId(), ['TITLE' => $newTitle], [])->isSuccess());
+        self::assertEquals($newTitle, $this->quoteService->get($addedItemResult->getId())->quote()->TITLE);
     }
 
     /**
@@ -127,28 +131,23 @@ class LeadTest extends TestCase
      */
     public function testCountByFilter(): void
     {
-        $before = $this->leadService->countByFilter();
+        $before = $this->quoteService->countByFilter();
 
-        $newItemsCount = 60;
+        $newItemsCount = 20;
         $items = [];
         for ($i = 1; $i <= $newItemsCount; $i++) {
             $items[] = ['TITLE' => 'TITLE-' . $i];
         }
 
         $cnt = 0;
-        foreach ($this->leadService->batch->add($items) as $item) {
+        foreach ($this->quoteService->batch->add($items) as $item) {
             $cnt++;
         }
 
         self::assertEquals(count($items), $cnt);
 
-        $after = $this->leadService->countByFilter();
+        $after = $this->quoteService->countByFilter();
 
         $this->assertEquals($before + $newItemsCount, $after);
-    }
-
-    protected function setUp(): void
-    {
-        $this->leadService = Fabric::getServiceBuilder()->getCRMScope()->lead();
     }
 }