+++ tests/Integration/Services/CRM/Quote/Service/QuoteUserfieldTest.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Lead\Service;
+namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service;
 
-use Bitrix24\SDK\Services\CRM\Lead\Service\LeadUserfield;
+use Bitrix24\SDK\Services\CRM\Quote\Service\QuoteUserfield;
 use Bitrix24\SDK\Tests\Builders\Services\CRM\Userfield\SystemUserfieldBuilder;
 use Bitrix24\SDK\Tests\Integration\Fabric;
 use Generator;
@@ -22,15 +22,20 @@ use PHPUnit\Framework\Attributes\CoversMethod;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
-#[CoversClass(LeadUserfield::class)]
-#[CoversMethod(LeadUserfield::class, 'add')]
-#[CoversMethod(LeadUserfield::class, 'get')]
-#[CoversMethod(LeadUserfield::class, 'list')]
-#[CoversMethod(LeadUserfield::class, 'delete')]
-#[CoversMethod(LeadUserfield::class, 'update')]
-class LeadUserfieldTest extends TestCase
+#[CoversClass(QuoteUserfield::class)]
+#[CoversMethod(QuoteUserfield::class, 'add')]
+#[CoversMethod(QuoteUserfield::class, 'get')]
+#[CoversMethod(QuoteUserfield::class, 'list')]
+#[CoversMethod(QuoteUserfield::class, 'delete')]
+#[CoversMethod(QuoteUserfield::class, 'update')]
+class QuoteUserfieldTest extends TestCase
 {
-    protected LeadUserfield $userfieldService;
+    protected QuoteUserfield $userfieldService;
+    
+    protected function setUp(): void
+    {
+        $this->userfieldService = Fabric::getServiceBuilder()->getCRMScope()->quoteUserfield();
+    }
 
     /**
      * @throws \Exception
@@ -64,22 +69,22 @@ class LeadUserfieldTest extends TestCase
     public function testGet(array $newUserFieldItem): void
     {
         $newUserfieldId = $this->userfieldService->add($newUserFieldItem)->getId();
-        $leadUserfieldItemResult = $this->userfieldService->get($newUserfieldId)->userfieldItem();
-        $this->assertEquals($newUserfieldId, $leadUserfieldItemResult->ID);
-        $this->assertEquals($newUserFieldItem['USER_TYPE_ID'], $leadUserfieldItemResult->USER_TYPE_ID);
-        $this->assertEquals('UF_CRM_' . $newUserFieldItem['FIELD_NAME'], $leadUserfieldItemResult->FIELD_NAME);
-        $this->assertEquals($newUserFieldItem['XML_ID'], $leadUserfieldItemResult->XML_ID);
+        $quoteUserfieldItemResult = $this->userfieldService->get($newUserfieldId)->userfieldItem();
+        $this->assertEquals($newUserfieldId, $quoteUserfieldItemResult->ID);
+        $this->assertEquals($newUserFieldItem['USER_TYPE_ID'], $quoteUserfieldItemResult->USER_TYPE_ID);
+        $this->assertEquals('UF_CRM_' . $newUserFieldItem['FIELD_NAME'], $quoteUserfieldItemResult->FIELD_NAME);
+        $this->assertEquals($newUserFieldItem['XML_ID'], $quoteUserfieldItemResult->XML_ID);
     }
 
     #[DataProvider('systemUserfieldsDemoDataDataProvider')]
     public function testUpdate(array $newUserFieldItem): void
     {
         $newUserfieldId = $this->userfieldService->add($newUserFieldItem)->getId();
-        $leadUserfieldItemResult = $this->userfieldService->get($newUserfieldId)->userfieldItem();
-        $this->assertEquals($newUserfieldId, $leadUserfieldItemResult->ID);
-        $this->assertEquals($newUserFieldItem['USER_TYPE_ID'], $leadUserfieldItemResult->USER_TYPE_ID);
-        $this->assertEquals('UF_CRM_' . $newUserFieldItem['FIELD_NAME'], $leadUserfieldItemResult->FIELD_NAME);
-        $this->assertEquals($newUserFieldItem['XML_ID'], $leadUserfieldItemResult->XML_ID);
+        $quoteUserfieldItemResult = $this->userfieldService->get($newUserfieldId)->userfieldItem();
+        $this->assertEquals($newUserfieldId, $quoteUserfieldItemResult->ID);
+        $this->assertEquals($newUserFieldItem['USER_TYPE_ID'], $quoteUserfieldItemResult->USER_TYPE_ID);
+        $this->assertEquals('UF_CRM_' . $newUserFieldItem['FIELD_NAME'], $quoteUserfieldItemResult->FIELD_NAME);
+        $this->assertEquals($newUserFieldItem['XML_ID'], $quoteUserfieldItemResult->XML_ID);
 
         $this->assertTrue(
             $this->userfieldService->update(
@@ -91,17 +96,13 @@ class LeadUserfieldTest extends TestCase
         );
 
         $ufFieldAfter = $this->userfieldService->get($newUserfieldId)->userfieldItem();
-        $this->assertEquals($leadUserfieldItemResult->EDIT_FORM_LABEL['en'] . 'QQQ', $ufFieldAfter->EDIT_FORM_LABEL['en']);
+        $this->assertEquals($quoteUserfieldItemResult->EDIT_FORM_LABEL['en'] . 'QQQ', $ufFieldAfter->EDIT_FORM_LABEL['en']);
     }
 
     public function testList(): void
     {
-        $leadUserfieldsResult = $this->userfieldService->list([], []);
-        $this->assertGreaterThanOrEqual(0, count($leadUserfieldsResult->getUserfields()));
+        $quoteUserfieldsResult = $this->userfieldService->list([], []);
+        $this->assertGreaterThanOrEqual(0, count($quoteUserfieldsResult->getUserfields()));
     }
 
-    protected function setUp(): void
-    {
-        $this->userfieldService = Fabric::getServiceBuilder()->getCRMScope()->leadUserfield();
-    }
 }