+++ tests/Integration/Services/CRM/Quote/Service/QuoteUserfieldUseCaseTest.php
@@ -11,60 +11,24 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Lead\Service;
+namespace Bitrix24\SDK\Tests\Integration\Services\CRM\Quote\Service;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Exceptions\TransportException;
-use Bitrix24\SDK\Services\CRM\Lead\Service\Lead;
-use Bitrix24\SDK\Services\CRM\Lead\Service\LeadUserfield;
+use Bitrix24\SDK\Services\CRM\Quote\Service\Quote;
+use Bitrix24\SDK\Services\CRM\Quote\Service\QuoteUserfield;
 use Bitrix24\SDK\Tests\Integration\Fabric;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Lead\Service\LeadUserfield::class)]
-class LeadUserfieldUseCaseTest extends TestCase
+#[\PHPUnit\Framework\Attributes\CoversClass(\Bitrix24\SDK\Services\CRM\Quote\Service\QuoteUserfield::class)]
+class QuoteUserfieldUseCaseTest extends TestCase
 {
-    protected Lead $leadService;
+    protected Quote $quoteService;
 
-    protected LeadUserfield $leadUserfieldService;
+    protected QuoteUserfield $quoteUserfieldService;
 
-    protected int $leadUserfieldId;
-
-    /**
-     * @throws BaseException
-     * @throws TransportException
-     */
-    public function testOperationsWithUserfieldFromLeadItem(): void
-    {
-        // get userfield metadata
-        $leadUserfieldItemResult = $this->leadUserfieldService->get($this->leadUserfieldId)->userfieldItem();
-        $ufOriginalFieldName = $leadUserfieldItemResult->getOriginalFieldName();
-        $ufFieldName = $leadUserfieldItemResult->FIELD_NAME;
-
-        // add lead with uf value
-        $fieldNameValue = 'test field value';
-        $newLeadId = $this->leadService->add(
-            [
-                'TITLE'      => 'test lead',
-                $ufFieldName => $fieldNameValue,
-            ]
-        )->getId();
-        $lead = $this->leadService->get($newLeadId)->lead();
-        $this->assertEquals($fieldNameValue, $lead->getUserfieldByFieldName($ufOriginalFieldName));
-
-        // update lead userfield value
-        $newUfValue = 'test 2';
-        $this->assertTrue(
-            $this->leadService->update(
-                $lead->ID,
-                [
-                    $ufFieldName => $newUfValue,
-                ]
-            )->isSuccess()
-        );
-        $leadItemResult = $this->leadService->get($lead->ID)->lead();
-        $this->assertEquals($newUfValue, $leadItemResult->getUserfieldByFieldName($ufOriginalFieldName));
-    }
-
+    protected int $quoteUserfieldId;
+    
     /**
      * @throws \Bitrix24\SDK\Services\CRM\Userfield\Exceptions\UserfieldNameIsTooLongException
      * @throws \Bitrix24\SDK\Core\Exceptions\TransportException
@@ -73,10 +37,10 @@ class LeadUserfieldUseCaseTest extends TestCase
      */
     protected function setUp(): void
     {
-        $this->leadService = Fabric::getServiceBuilder()->getCRMScope()->lead();
-        $this->leadUserfieldService = Fabric::getServiceBuilder()->getCRMScope()->leadUserfield();
+        $this->quoteService = Fabric::getServiceBuilder()->getCRMScope()->quote();
+        $this->quoteUserfieldService = Fabric::getServiceBuilder()->getCRMScope()->quoteUserfield();
 
-        $this->leadUserfieldId = $this->leadUserfieldService->add(
+        $this->quoteUserfieldId = $this->quoteUserfieldService->add(
             [
                 'FIELD_NAME'        => sprintf('%s%s', substr((string)random_int(0, PHP_INT_MAX), 0, 3), time()),
                 'EDIT_FORM_LABEL'   => [
@@ -98,6 +62,43 @@ class LeadUserfieldUseCaseTest extends TestCase
 
     protected function tearDown(): void
     {
-        $this->leadUserfieldService->delete($this->leadUserfieldId);
+        $this->quoteUserfieldService->delete($this->quoteUserfieldId);
     }
+
+    /**
+     * @throws BaseException
+     * @throws TransportException
+     */
+    public function testOperationsWithUserfieldFromQuoteItem(): void
+    {
+        // get userfield metadata
+        $quoteUserfieldItemResult = $this->quoteUserfieldService->get($this->quoteUserfieldId)->userfieldItem();
+        $ufOriginalFieldName = $quoteUserfieldItemResult->getOriginalFieldName();
+        $ufFieldName = $quoteUserfieldItemResult->FIELD_NAME;
+
+        // add quote with uf value
+        $fieldNameValue = 'test field value';
+        $newQuoteId = $this->quoteService->add(
+            [
+                'TITLE'      => 'test quote',
+                $ufFieldName => $fieldNameValue,
+            ]
+        )->getId();
+        $quote = $this->quoteService->get($newQuoteId)->quote();
+        $this->assertEquals($fieldNameValue, $quote->getUserfieldByFieldName($ufOriginalFieldName));
+
+        // update quote userfield value
+        $newUfValue = 'test 2';
+        $this->assertTrue(
+            $this->quoteService->update(
+                $quote->ID,
+                [
+                    $ufFieldName => $newUfValue,
+                ]
+            )->isSuccess()
+        );
+        $quoteItemResult = $this->quoteService->get($quote->ID)->quote();
+        $this->assertEquals($newUfValue, $quoteItemResult->getUserfieldByFieldName($ufOriginalFieldName));
+    }
+    
 }