+++ CHANGELOG.md
@@ -4,6 +4,39 @@
 
 ### Added
 
+- Added service `Services\CRM\Quote\Service\Quote` with support methods,
+  see [crm.quote.* methods](https://github.com/bitrix24/b24phpsdk/issues/179):
+    - `fields` returns a list of fields for the quote
+    - `get` returns the settings of the quote by Id
+    - `list` returns a list of quote
+    - `add` creates a new quote
+    - `delete` deletes a quote
+    - `update` modifies the quote
+    - `countByFilter` count quotes by filter
+- Added support for events:
+    - `OnCrmQuoteAdd`
+    - `OnCrmQuoteDelete`
+    - `OnCrmQuoteUpdate`
+    - `OnCrmQuoteUserFieldAdd`
+    - `OnCrmQuoteUserFieldDelete`
+    - `OnCrmQuoteUserFieldSetEnumValues`
+    - `OnCrmQuoteUserFieldUpdate`
+- Added service `Services\CRM\Quote\Service\QuoteUserfield` with support methods:
+    - `add` add userfield to a quote
+    - `get` get userfield to a quote
+    - `list` list userfields
+    - `delete` delete userfield
+    - `update` update userfield
+- Added service `Services\CRM\Quote\Service\QuoteProductRows` with support methods:
+    - `set` Adds products to a quote
+    - `get` Returns the products of a quote
+- Added service `Services\CRM\Quote\Service\QuoteContact` with support methods,
+    - `fields` get fiels for quote contact connection
+    - `setItems` set contacts related with quote
+    - `get` get contacts related to quote
+    - `deleteItems` delete all relations for quote
+    - `add` add contact relation with quote
+    - `delete` delete contact relation with quote
 - Added service `Services\CRM\Deal\Service\DealRecurring` with support methods,
   see [crm.deal.recurring.* methods](https://github.com/bitrix24/b24phpsdk/issues/160):
     - `fields` returns a list of fields for the recurring deal template